requirements.txt
qrisp
networkx
matplotlib
Pillow
requests
